+++ src/base/fixed_data_block.h
@@ -22,16 +22,16 @@ public:
     /// @param sample_rate_per_unit as a data_block must not represent data which
     ///                             is associated to time, the sample_rate is given
     ///                             in "per unit" (e.g. "s" or "hz", etc.)
-    FixedDataBlock (QSharedPointer<QVector<float32> > data, float32 sample_rate_per_unit);
+    FixedDataBlock (QSharedPointer<QVector<float32> > data, float64 sample_rate_per_unit);
 
     //-------------------------------------------------------------------------
     virtual ~FixedDataBlock () {}
 
     //-------------------------------------------------------------------------
-    virtual QSharedPointer<DataBlock> createSubBlock (uint32 start, uint32 length) const;
+    virtual QSharedPointer<DataBlock> createSubBlock (size_t start, size_t length) const;
 
     //-------------------------------------------------------------------------
-    virtual float32 const& operator[] (uint32 index) const;
+    virtual float32 const& operator[] (size_t index) const;
 
     //-------------------------------------------------------------------------
     virtual float32 getMin () const;
@@ -56,16 +56,16 @@ private:
     Q_DISABLE_COPY (FixedDataBlock);
 
     //---------------------------------------------------------------------------------------------
-    FixedDataBlock (FixedDataBlock const& base, unsigned new_start, unsigned new_length);
+    FixedDataBlock (FixedDataBlock const& base, size_t new_start, size_t new_length);
 
     //---------------------------------------------------------------------------------------------
     static QSharedPointer<DataBlock> calculateStandardDeviationImpl (std::list<QSharedPointer<DataBlock const> > const &data_blocks,
                                                                       QSharedPointer<DataBlock> means);
 
     QSharedPointer<QVector<float32> > data_;
-    unsigned start_index_;
+    size_t start_index_;
 
-    static unsigned instance_count_;
+    static size_t instance_count_;
 };
 
 }
