+++ src/file_handling/basic_header.cpp
@@ -6,13 +6,13 @@ namespace SigViewer_
 {
 
 //-----------------------------------------------------------------------------
-QString BasicHeader::getFileTypeString () const
+QString BasicHeader::getFileTypeString() const
 {
     return file_type_string_;
 }
 
 //-----------------------------------------------------------------------------
-float32 BasicHeader::getSampleRate () const
+float64 BasicHeader::getSampleRate() const
 {
     return sample_rate_;
 }
@@ -23,7 +23,7 @@ uint32 BasicHeader::getNumberEvents() const
     return number_events_;
 }
 
-void BasicHeader::setNumberEvents (uint32 number_events)
+void BasicHeader::setNumberEvents(uint32 number_events)
 {
     number_events_ = number_events;
 }
@@ -34,13 +34,13 @@ double BasicHeader::getEventSamplerate() const
     return sample_rate_;
 }
 
-void BasicHeader::setEventSamplerate (double event_sample_rate)
+void BasicHeader::setEventSamplerate(double event_sample_rate)
 {
     event_sample_rate_ = event_sample_rate;
 }
 
 //-------------------------------------------------------------------------
-QSharedPointer<SignalChannel const> BasicHeader::getChannel (ChannelID id) const
+QSharedPointer<SignalChannel const> BasicHeader::getChannel(ChannelID id) const
 {
     if (channels_.contains(id))
         return channels_[id];
@@ -55,19 +55,19 @@ unsigned BasicHeader::getNumberChannels() const
 }
 
 //-------------------------------------------------------------------------
-void BasicHeader::setFileTypeString (QString const& file_type_string)
+void BasicHeader::setFileTypeString(QString const& file_type_string)
 {
     file_type_string_ = file_type_string;
 }
 
 //-------------------------------------------------------------------------
-void BasicHeader::setSampleRate (float sample_rate)
+void BasicHeader::setSampleRate(float64 sample_rate)
 {
     sample_rate_ = sample_rate;
 }
 
 //-------------------------------------------------------------------------
-void BasicHeader::addChannel (ChannelID id, QSharedPointer<SignalChannel const> channel)
+void BasicHeader::addChannel(ChannelID id, QSharedPointer<SignalChannel const> channel)
 {
     channels_[id] = channel;
 }
